requirements.txt
openai-whisper
anthropic
ffmpeg-python
python-dotenv
